+++ src/main/java/wooteco/subway/domain/element/Section.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.element;
 
 import java.util.List;
 import java.util.Objects;
