+++ src/main/java/wooteco/subway/repository/StationRepository.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Repository;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.dao.entity.StationEntity;
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.element.Station;
 import wooteco.subway.exception.NotFoundException;
 
 @Repository
