+++ src/main/java/wooteco/subway/service/LineService.java
@@ -3,10 +3,10 @@ package wooteco.subway.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import wooteco.subway.domain.Line;
-import wooteco.subway.domain.Section;
-import wooteco.subway.domain.Sections;
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.element.Line;
+import wooteco.subway.domain.element.Section;
+import wooteco.subway.domain.element.Sections;
+import wooteco.subway.domain.element.Station;
 import wooteco.subway.exception.BadRequestException;
 import wooteco.subway.repository.LineRepository;
 import wooteco.subway.repository.SectionRepository;
