+++ src/main/java/wooteco/subway/service/PathService.java
@@ -3,9 +3,9 @@ package wooteco.subway.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import wooteco.subway.domain.Section;
-import wooteco.subway.domain.Station;
-import wooteco.subway.domain.SubwayGraph;
+import wooteco.subway.domain.element.Section;
+import wooteco.subway.domain.element.Station;
+import wooteco.subway.domain.path.SubwayGraph;
 import wooteco.subway.domain.fare.Fare;
 import wooteco.subway.repository.SectionRepository;
 import wooteco.subway.repository.StationRepository;
