+++ src/main/java/wooteco/subway/service/StationService.java
@@ -3,7 +3,7 @@ package wooteco.subway.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.element.Station;
 import wooteco.subway.exception.BadRequestException;
 import wooteco.subway.repository.StationRepository;
 import wooteco.subway.service.dto.request.StationRequest;
