+++ src/test/java/wooteco/subway/dao/SectionDaoTest.java
@@ -13,7 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import wooteco.subway.dao.entity.SectionEntity;
-import wooteco.subway.domain.Line;
+import wooteco.subway.domain.element.Line;
 
 @JdbcTest
 @Import(SectionDao.class)
