+++ src/test/java/wooteco/subway/domain/SectionsTest.java
@@ -13,6 +13,9 @@ import static wooteco.subway.domain.fixtures.TestFixtures.합정;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.element.Line;
+import wooteco.subway.domain.element.Section;
+import wooteco.subway.domain.element.Sections;
 
 class SectionsTest {
