+++ src/test/java/wooteco/subway/domain/fare/AgePolicyTest.java
@@ -1,10 +1,10 @@
 package wooteco.subway.domain.fare;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.policy.AgePolicy;
 
 class AgePolicyTest {
