+++ src/test/java/wooteco/subway/domain/fare/FareTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.policy.AgePolicy;
 
 class FareTest {
     private final Fare fare = new Fare();
